+++ src/components/Common/index.js
@@ -1,18 +1,35 @@
 import React, { Component } from 'react'
-import {Text, View, TouchableOpacity } from 'react-native'
+import {Text, TextInput, View, TouchableOpacity, KeyboardAvoidingView } from 'react-native'
 
 import { styles } from './styles'
 
 export class Container extends Component{
 	render(){
 
-		let {addStyles, children} = this.props
-
+		let {formContainer, addStyles, children} = this.props
+		
+		if (formContainer) {
+			return (
+				<KeyboardAvoidingView behavior='padding' style={[styles.container, addStyles]}>
+					{children}
+				</KeyboardAvoidingView>
+			)
+		}
+		
 		return (
 			<View style={[styles.container, addStyles]}>{children}</View>
 		)
 	}
 
+}
+
+export class Input extends Component{
+
+	render(){
+		return(
+			<TextInput {...this.props} style={[styles.input, this.props.addStyles]} />
+		)
+	}
 }
 
 export const Button = ({text, handler, addStyles}) => {
@@ -38,14 +55,6 @@ export const Title = ({text, isSubtitle, addStyles}) => {
 	return (
 		<Text style={[style, addStyles]}>{text}</Text>
 	)
-
-}
-
-export const Input = ({}) => {
-	
-	return {
-
-	}
 
 }