+++ src/components/Common/styles.js
@@ -4,7 +4,7 @@ import { StyleSheet } from 'react-native'
 import { common } from '../../styles'
 
 export const styles = StyleSheet.create({
-  container: common.centeredContainer,
+  container: common.container,
   title: common.title,
   subtitle: common.subtitle,
   btn: common.btn,