+++ src/components/Quiz/styles.js
@@ -4,5 +4,5 @@ import { StyleSheet } from 'react-native'
 import { common } from '../../styles'
 
 export const styles = StyleSheet.create({
-  noStyles:{}
+  centered : common.centeredContainer
 });