+++ src/components/QuizForm/styles.js
@@ -1,13 +1,13 @@
 
 import { StyleSheet } from 'react-native'
 
-import { common } from '../../styles'
+import { space } from '../../styles'
 
 export const styles = StyleSheet.create({
-  container: common.container,
-  title: common.title,
-  subtitle: common.subtitle,
-  btn: common.btn,
-  btnText: common.btnText,
-  input: common.input
+	addHeight : {
+		height: space.xxlrg
+	},
+	addMargin: {
+		marginTop: space.xlrg
+	}
 });