+++ src/styles/index.js
@@ -25,14 +25,15 @@ export const space ={
 	med : 20,
 	lrg : 30,
 	xlrg : 50,
+	xxlrg: 75
 }
 
 export const common ={
 	container: {
 	    flex: 1,
 	    backgroundColor: colors.white,
-	    marginLeft: space.small,
-	    marginRight: space.small,
+	    paddingLeft: space.small,
+	    paddingRight: space.small,
 	    alignItems: 'stretch',
 	    justifyContent: 'center',
 	},
@@ -70,6 +71,7 @@ export const common ={
 	    borderRadius: space.xxsmall,
 	    borderWidth: space.xxxsmall,
 	    borderColor: colors.darkGray,
+	    alignItems: 'stretch',
 	}
 }