+++ src/views/AddCard.js
@@ -2,10 +2,16 @@ import React, { Component } from 'react'
 import { FlatList } from 'react-native'
 import { connect } from 'react-redux'
 
-import QuizItem from '../components/QuizItem'
+// import Quiz from '../components/Quiz'
 
-class QuizList extends Component {
+class AddCard extends Component {
 	
+	static navigationOptions = () => {
+		return {
+			title: 'Add a card'
+		}
+	}
+
 	constructor(props) {
 		super(props)
 		this.renderQuiz = this.renderQuiz.bind(this);
@@ -16,12 +22,13 @@ class QuizList extends Component {
 		nav.navigate('QuizDetail',{quiz})
 	}
 
-	renderQuiz({item, key}){
+	renderCard({item, key}){
 		// console.log('QUIZ WITHIN LOOP', item.title)
+		/*
 		return (
-			<QuizItem id={key} quiz={item} nav={() => this.goToDeck(this.props.navigation, item)}/>
+			<Quiz id={key} quiz={item} nav={() => this.goToDeck(this.props.navigation, item)}/>
 
-		)
+		) */
 	}
 
 	render(){
@@ -43,4 +50,4 @@ class QuizList extends Component {
 function mapStateToProps({quizList}){
 	return {quizList}
 }
-export default connect(mapStateToProps)(QuizList);
+export default connect(mapStateToProps)(AddCard);